+++ include/decision_tree_regressor.hpp
@@ -144,7 +144,7 @@ private:
     void cost_complexity_prune();
 public:
     explicit DecisionTreeRegressor(int max_depth = 32,
-                          int min_samples_split = 5,
+                          size_t min_samples_split = 5,
                           size_t min_samples_leaf = 2,
                           const std::string& string_criterion = "mse",
                           double ccp_alpha = 0.0);
